pyproject.toml
[tool.poetry]
name = "carscanner"
version = "0.1.0"
description = "Car offer scanner"
authors = ["Rafal Krupinski <rkrupinski+github@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
mattes-allegro-pl = {path = '../mattes-allegro-pl'}
CherryPy = "^18.1"
PyYAML = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
